+++ src/termvisage/tui/keys.py
@@ -15,20 +15,20 @@ from term_image.image import GraphicsImage
 from term_image.utils import get_cell_size, get_terminal_size
 
 from .. import __version__, logging
-from ..config import context_keys, expand_key
+from ..config import config_options, context_keys, expand_key
 from . import main
 from .render import resync_grid_rendering
 from .widgets import (
     Image,
     ImageCanvas,
-    bottom_bar,
+    action_bar,
     confirmation,
     confirmation_overlay,
     expand,
+    footer,
     image_box,
     image_grid,
     image_grid_box,
-    key_bar,
     main as main_widget,
     menu,
     menu_box,
@@ -47,7 +47,8 @@ def disable_actions(context: str, *actions: str) -> None:
     for action in actions:
         keyset[action][4] = False
         keys[context][keyset[action][0]][1] = False
-        display_context_keys(context)
+    if context == main.get_context() or context == "global":
+        action_bar.update(context)
 
 
 def enable_actions(context: str, *actions: str) -> None:
@@ -55,14 +56,14 @@ def enable_actions(context: str, *actions: str) -> None:
     for action in actions:
         keyset[action][4] = True
         keys[context][keyset[action][0]][1] = True
-        display_context_keys(context)
+    if context == main.get_context() or context == "global":
+        action_bar.update(context)
 
 
 def hide_actions(context: str, *actions: str) -> None:
     keyset = context_keys[context]
     for action in actions:
         keyset[action][3] = False
-        display_context_keys(context)
     disable_actions(context, *actions)
 
 
@@ -70,7 +71,6 @@ def show_actions(context: str, *actions: str) -> None:
     keyset = context_keys[context]
     for action in actions:
         keyset[action][3] = True
-        display_context_keys(context)
     enable_actions(context, *actions)
 
 
@@ -225,32 +225,6 @@ def display_context_help(context: str) -> None:
     )
 
 
-def display_context_keys(context: str) -> None:
-    """Updates the Key/Action bar with the actions in the given context.
-
-    Includes "global" actions for all contexts except those in `no_globals`.
-    """
-    actions = (
-        *context_keys[context].items(),
-        *(() if context in no_globals else context_keys["global"].items()),
-    )
-
-    # The underscores and blocks (U+2588) are to prevent wrapping amidst keys
-    key_bar.original_widget.set_text(
-        [
-            [
-                ("key" if enabled else "disabled key", action.replace(" ", "\u2800")),
-                ("key", "\u2800"),
-                ("key" if enabled else "disabled key", f"[{symbol}]"),
-                " ",
-            ]
-            for action, (_, symbol, _, visible, enabled) in actions
-            if visible
-        ]
-    )
-    adjust_bottom_bar()
-
-
 def register_key(*args: Tuple[str, str]) -> FunctionType:
     """Returns a decorator to register a function to some context action(s).
 
@@ -316,6 +290,19 @@ def set_confirmation(
     getattr(main.ImageClass, "clear", lambda: True)()
 
 
+def update_footer_expand_collapse_icon():
+    if not config_options.show_footer:
+        return
+
+    expand.set_text(
+        [
+            "\u25B2" if action_bar._ti_collapsed else "\u25BC",
+            " ",
+            ("key", f" {expand_key[1]} "),
+        ]
+    )
+
+
 # Context Actions
 
 # {<context>: {<key>: [<func>, <state>], ...}, ...}
@@ -329,21 +316,18 @@ def quit():
     raise urwid.ExitMainLoop()
 
 
-@register_key(("global", "Key Bar"))
+@register_key(("global", "Expand/Collapse Footer"))
 def expand_collapse_keys():
     if expand._ti_shown:
-        if key_bar._ti_collapsed and key_bar_rows() > 1:
-            expand.original_widget.set_text(f"\u25BC [{expand_key[1]}]")
-            main_widget.contents[-1] = (
-                bottom_bar,
-                ("given", key_bar_rows()),
-            )
-            key_bar._ti_collapsed = False
+        if action_bar._ti_collapsed and action_bar_rows() > 1:
+            update_footer_expand_collapse_icon()
+            main_widget.contents[-1] = (footer, ("given", action_bar_rows()))
+            action_bar._ti_collapsed = False
             getattr(main.ImageClass, "clear", lambda: True)() or ImageCanvas.change()
-        elif not key_bar._ti_collapsed:
-            expand.original_widget.set_text(f"\u25B2 [{expand_key[1]}]")
-            main_widget.contents[-1] = (bottom_bar, ("given", 1))
-            key_bar._ti_collapsed = True
+        elif not action_bar._ti_collapsed:
+            update_footer_expand_collapse_icon()
+            main_widget.contents[-1] = (footer, ("given", 1))
+            action_bar._ti_collapsed = True
             getattr(main.ImageClass, "clear", lambda: True)() or ImageCanvas.change()
 
 
@@ -353,36 +337,32 @@ def help():
     getattr(main.ImageClass, "clear", lambda: True)()
 
 
-def adjust_bottom_bar():
-    cols = get_terminal_size()[0]
-    rows = key_bar.original_widget.rows((cols,))
+def adjust_footer():
+    if not config_options.show_footer:
+        return
+
+    needed_rows = action_bar.rows((get_terminal_size()[0],))
     if expand._ti_shown:
-        if rows == 1:
-            bottom_bar.contents.pop()
+        if needed_rows == 1:
+            footer.contents.pop()
             expand._ti_shown = False
-    elif rows > 1:
-        bottom_bar.contents.append(
-            (expand, ("given", len(expand.original_widget.text), False))
-        )
+    elif needed_rows > 1:
+        footer.contents.append((expand, ("pack", None, False)))
         expand._ti_shown = True
 
-    if not key_bar._ti_collapsed:
-        new_rows = key_bar_rows()
-        if main_widget.contents[-1][1][1] != new_rows:
+    if not action_bar._ti_collapsed:
+        if main_widget.contents[-1][1][1] != (rows := action_bar_rows()):
+            main_widget.contents[-1] = (footer, ("given", rows))
             getattr(main.ImageClass, "clear", lambda: True)()
-        main_widget.contents[-1] = (
-            bottom_bar,
-            ("given", new_rows),
-        )
 
 
-def key_bar_rows():
+def action_bar_cols():
     # Consider columns occupied by the expand key and the divider
-    cols = (
-        get_terminal_size()[0]
-        - (len(expand.original_widget.text) + 2) * expand._ti_shown
-    )
-    return key_bar.original_widget.rows((cols,))
+    return get_terminal_size()[0] - (expand.pack()[0] + 2) * expand._ti_shown
+
+
+def action_bar_rows():
+    return action_bar.rows((action_bar_cols(),))
 
 
 def resize():
@@ -405,7 +385,7 @@ def resize():
             if main.grid_active.is_set():
                 resync_grid_rendering()
 
-    adjust_bottom_bar()
+    adjust_footer()
     getattr(main.ImageClass, "clear", lambda: True)() or ImageCanvas.change()
 
 
@@ -733,7 +713,7 @@ def close():
 
 logger = _logging.getLogger(__name__)
 no_globals = {"global", "confirmation", "full-grid-image", "overlay"}
-key_bar._ti_collapsed = True
+action_bar._ti_collapsed = True
 expand._ti_shown = True
 
 # Used in the "confirmation" context.